Working on: RIDES, shared memory - populate struct
